backend/requirements.txt
fastapi
uvicorn[standard]
httpx
numpy
scikit-learn
xgboost
pydantic
python-dotenv
orjson
google-generativeai
cachetools
# Optional: for ML model features. Install separately if needed:
# pip install "tensorflow>=2.12"
